+++ MK/module/nextion/Nextion.h
@@ -21,7 +21,7 @@
 #include "NexConfig.h"
 #include "NexTouch.h"
 #include "NexHardware.h"
-#include "NexDownload.h"
+#include "NexUpload.h"
 
 #include "NexButton.h"
 //#include "NexCrop.h"